requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
ase==3.26.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bibtexparser==1.4.3
bleach==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chex==0.1.89
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.3
cuequivariance==0.6.1
cuequivariance-ops-cu12==0.6.1
cuequivariance-ops-torch-cu12==0.6.1
cuequivariance-torch==0.6.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dm-haiku==0.0.15
e3nn==0.4.4
etils==1.13.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.13.1
flax==0.10.6
fonttools==4.60.0
fqdn==1.5.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jax==0.5.3
jax-cuda12-pjrt==0.5.3
jax-cuda12-plugin==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.4
jmp==0.0.4
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lightning-utilities==0.15.2
lmdb==1.7.3
mace-torch==0.3.14
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
matscipy==1.1.1
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.3
monty==2025.3.3
mpmath==1.3.0
msgpack==1.1.1
narwhals==2.5.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.13
orjson==3.11.3
packaging==25.0
palettable==3.3.3
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.4.0
plotly==6.3.0
prettytable==3.16.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
pymatgen==2025.6.14
pynvml==13.0.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python_hostlist==2.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scipy==1.16.2
Send2Trash==1.8.3
setuptools==70.2.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spglib==2.6.0
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorstore==0.1.77
terminado==0.18.1
tinycss2==1.4.0
toolz==1.0.0
torch==2.8.0+cu128
torch-ema==0.3
torchmetrics==1.8.2
torchvision==0.23.0+cu128
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.10
triton==3.4.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
uncertainties==3.2.3
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
zipp==3.23.0
